Changed x-axis to show actual time. Made GUI responsive even for slow time frequencies.

+++ qbpm_monitor.py
@@ -12,6 +12,7 @@ import pyqtgraph as pg
 import tango
 import numpy
 import time
+import datetime
 
 
 class QbpmMonitor(QtGui.QWidget):
@@ -22,7 +23,8 @@ class QbpmMonitor(QtGui.QWidget):
         self.posz_target = 0
         self.avgcurr_target = 0
         self.qbpm = qbpm(address, distance)
-        self.frequency = 5.0    # in Hz
+        self.qbpm.frequency = 5.0  # in Hz
+        self.qbpm.backlog = 120  # in s
         self.polling = False
         self._generator_poll = None
         self._timerId_poll = None
@@ -33,6 +35,7 @@ class QbpmMonitor(QtGui.QWidget):
         self.dcm_bragg_tserver = tango.DeviceProxy('hzgpp05vme0:10000/dcm_bragg')
         self.dcm_bragg_angle = self.dcm_bragg_tserver.Position
         self.dcm_pitch_tserver = tango.DeviceProxy('hzgpp05vme0:10000/dcm_xtal2_pitch')
+        self.heartbeat = time.time()
 
         self.initUI()
 
@@ -41,7 +44,7 @@ class QbpmMonitor(QtGui.QWidget):
         # labels
         self.poll_label = QtGui.QLabel("poll")
         self.feedback_label = QtGui.QLabel("feedback")
-        self.ll_label = QtGui.QLabel("log length")
+        self.ll_label = QtGui.QLabel("backlog (s)")
         self.freq_label = QtGui.QLabel("frequency")
         self.pitch_label = QtGui.QLabel("DCM pitch: {}".format(self.dcm_pitch_tserver.Position))
         # quit button
@@ -62,13 +65,13 @@ class QbpmMonitor(QtGui.QWidget):
         self.fbtn.clicked.connect(self.toggle_feedback)
         self.fbtn.resize(qbtn.sizeHint())
         self.fbtn.setIcon(self.style().standardIcon(QtWidgets.QStyle.SP_MediaPlay))
-        # log length text field
-        self.lltext = QtGui.QLineEdit(str(self.qbpm.log_length))
+        # backlog text field
+        self.lltext = QtGui.QLineEdit(str(self.qbpm.backlog))
         self.lltext.setValidator(QtGui.QIntValidator())
         self.lltext.setMaxLength(6)
-        self.lltext.returnPressed.connect(self.change_loglength)
+        self.lltext.returnPressed.connect(self.change_backlog)
         # frequency text field
-        self.ftext = QtGui.QLineEdit(str(self.frequency))
+        self.ftext = QtGui.QLineEdit(str(self.qbpm.frequency))
         self.ftext.setValidator(QtGui.QDoubleValidator())
         self.ftext.setMaxLength(6)
         self.ftext.returnPressed.connect(self.change_frequency)
@@ -85,11 +88,15 @@ class QbpmMonitor(QtGui.QWidget):
         font.setPixelSize(16)
         # make PlotWidgets
         self.plot_main = pg.GraphicsLayoutWidget()
-        self.plot_avgcurr = self.plot_main.addPlot(title='avg. current', row=0, col=0)
-        self.plot_petracurrent = self.plot_main.addPlot(title='PETRA beam current', row=0, col=1)
+        self.avgcurr_timeaxis = TimeAxisItem(orientation='bottom')
+        self.plot_avgcurr = self.plot_main.addPlot(title='avg. current', row=0, col=0, axisItems={'bottom': self.avgcurr_timeaxis})
+        self.petracurr_timeaxis = TimeAxisItem(orientation='bottom')
+        self.plot_petracurrent = self.plot_main.addPlot(title='PETRA beam current', row=0, col=1, axisItems={'bottom': self.petracurr_timeaxis})
 #        self.plot_main.nextRow()
-        self.plot_posx = self.plot_main.addPlot(title='x-position', row=1, col=0)
-        self.plot_posz = self.plot_main.addPlot(title='z-position', row=1, col=1)
+        self.posx_timeaxis = TimeAxisItem(orientation='bottom')
+        self.plot_posx = self.plot_main.addPlot(title='x-position', row=1, col=0, axisItems={'bottom': self.posx_timeaxis})
+        self.posy_timeaxis = TimeAxisItem(orientation='bottom')
+        self.plot_posz = self.plot_main.addPlot(title='z-position', row=1, col=1, axisItems={'bottom': self.posy_timeaxis})
         # assign qbpm data tp styles to PlotWidgets
         styles = {'avgcurr_log': (self.plot_avgcurr, log_pen),
                   'avgcurr_mvavg_log': (self.plot_avgcurr, avg_pen),
@@ -104,6 +111,7 @@ class QbpmMonitor(QtGui.QWidget):
                   }
         # plot curves
         for key, style in styles.items():
+            #self.curves[key] = style[0].plot(self.qbpm.log_arrays[key], pen=style[1], symbol='o')
             self.curves[key] = style[0].plot(self.qbpm.log_arrays[key], pen=style[1])
             style[0].getAxis("bottom").tickFont = font
             style[0].getAxis("bottom").setStyle(tickTextOffset=20)
@@ -117,7 +125,7 @@ class QbpmMonitor(QtGui.QWidget):
         layout = QtGui.QGridLayout()
         self.setLayout(layout)
 
-        ## Add widgets to the layout in their proper positions
+        # Add widgets to the layout in their proper positions
         layout.addWidget(self.poll_label, 0, 0)
         layout.addWidget(self.feedback_label, 1, 0)
         layout.addWidget(self.ll_label, 3, 0)
@@ -135,14 +143,14 @@ class QbpmMonitor(QtGui.QWidget):
         layout.setColumnStretch(1,0.1)
         layout.setColumnStretch(2,1)
 
-        ## Display the widget as a new window
+        # Display the widget as a new window
         self.setWindowTitle(self.title)
         self.show()
 
     def plot_update(self):
         for key, names in self.qbpm.log_names.items():
             for name in names:
-                self.curves[name].setData(self.qbpm.log_arrays[name])
+                self.curves[name].setData(self.qbpm.log_xarr, self.qbpm.log_arrays[name])
 
     def reset_logs(self):
         self.qbpm.reset_logs()
@@ -171,6 +179,7 @@ class QbpmMonitor(QtGui.QWidget):
         self._generator_poll = None
         self._timerId_poll = None
         self.rbtn.setIcon(self.style().standardIcon(QtWidgets.QStyle.SP_MediaPlay))
+        self.heartbeat = time.time()
 
     def toggle_feedback(self):
         self.feedback = not self.feedback
@@ -232,43 +241,49 @@ class QbpmMonitor(QtGui.QWidget):
 
     def timerEvent(self, event):
         # This is called every time the GUI is idle.
-        time.sleep(1/self.frequency)
-        if self._generator_poll is None:
-            if self._generator_feedback is not None:
-                print('timer.')
-                self.stop_loop_feedback()
-            return
-        try:
-            next(self._generator_poll)  # Run the next iteration
-            if self._generator_feedback is None:
+        if self.check_pulse():
+            if self._generator_poll is None:
+                if self._generator_feedback is not None:
+                    print('timer.')
+                    self.stop_loop_feedback()
                 return
             try:
-                next(self._generator_feedback)
-            except:
-                print('generator feedback exception')
-                self.stop_loop_feedback()
-        except StopIteration:
-            print('generator poll exception')
-            self.stop_loop_feedback()  # Iteration has finshed, kill the timer
-            self.stop_loop_poll()  # Iteration has finshed, kill the timer
+                next(self._generator_poll)  # Run the next iteration
+                if self._generator_feedback is None:
+                    return
+                try:
+                    next(self._generator_feedback)
+                except:
+                    print('generator feedback exception')
+                    self.stop_loop_feedback()
+            except StopIteration:
+                print('generator poll exception')
+                self.stop_loop_feedback()  # Iteration has finshed, kill the timer
+                self.stop_loop_poll()  # Iteration has finshed, kill the timer
 
-    def change_loglength(self):
+    def check_pulse(self):
+        timedelta = time.time() - self.heartbeat
+        update_delay = 1/self.qbpm.frequency
+        time_to_update = False
+        if timedelta > update_delay:
+            time_to_update = True
+            self.heartbeat = time.time()
+        return time_to_update
+
+    def change_backlog(self):
         if not self.lltext.text():
             return
-        log_length = int(self.lltext.text())
-        if log_length < self.qbpm.box_length:
-            log_length = self.qbpm.box_length
-            self.lltext.setText(str(log_length))
-        self.qbpm.change_log_length(log_length)
+        backlog = int(self.lltext.text())
+        self.qbpm.change_backlog(backlog)
+        self.lltext.setText(str(self.qbpm.backlog))
 
     def change_frequency(self):
         if not self.ftext.text():
             return
         frequency = float(self.ftext.text())
-        if frequency > 10:
-            frequency = 10.0
-            self.ftext.setText(str(frequency))
-        self.frequency = frequency
+        self.qbpm.change_frequency(frequency)
+        self.ftext.setText(str(self.qbpm.frequency))
+
 
 
 class qbpm():
@@ -276,11 +291,16 @@ class qbpm():
         self.tserver = tango.DeviceProxy(address)
         self.distance = distance
         self.petra = tango.DeviceProxy('hzgpp05vme1:10000/PETRA/GLOBALS/keyword')
-        self.log_length = 400
+        self.frequency = 5  # in Hz
+        self.backlog = 120  # in s
+        self.log_length = self.calc_log_length(self.backlog, self.frequency)
         self.log_names = {'log_vals': ['posx_log', 'posz_log', 'avgcurr_log', 'petracurrent_log'],
                           'log_mvavg': ['posx_mvavg_log', 'posz_mvavg_log', 'avgcurr_mvavg_log'],
-                          'log_target': ['posx_target_log', 'posz_target_log', 'avgcurr_target_log']}
-        self.reset_logs()
+                          'log_target': ['posx_target_log', 'posz_target_log', 'avgcurr_target_log'],
+                          }
+        self.log_xarr = numpy.zeros(self.log_length)
+        self.adept_log_xarr()
+        self.reset_logs()  # initialize log_arrays with appropriate log_length
         self.box_length = 40
         self.posx_target = 0
         self.posz_target = 0
@@ -292,6 +312,7 @@ class qbpm():
         for key, names in self.log_names.items():
             for name in names:
                 self.log_arrays[name][:] = numpy.roll(self.log_arrays[name], -1)
+        self.log_xarr[:] = numpy.roll(self.log_xarr, -1)
         # query qbpm and petra current, append to log array
         try:
             bc = self.petra.BeamCurrent
@@ -314,20 +335,54 @@ class qbpm():
                 self.log_arrays[key][-1] = targets[n]
             else:
                 self.log_arrays[key][-1] = self.log_arrays[self.log_names['log_mvavg'][n]][-1]
+        # finally append unix timestamp to log_xarr
+        self.log_xarr[-1] = self.timestamp()
+        print(self.log_xarr[-1] - self.log_xarr[0])
 
     def change_log_length(self, log_length):
         len_diff = abs(self.log_length - log_length)
         if log_length > self.log_length:
             for key, names in self.log_names.items():
                 for name in names:
-                    self.log_arrays[name].resize(log_length)
-                    self.log_arrays[name][:] = numpy.roll(self.log_arrays[name], len_diff)
+                    tmparr = numpy.full(log_length, self.log_arrays[name][0])
+                    tmparr[-self.log_arrays[name].size:] = self.log_arrays[name]
+                    self.log_arrays[name] = tmparr
+            tmparr = numpy.resize(self.log_xarr, log_length)
+            self.log_xarr = tmparr
+            self.adept_log_xarr()
         else:
             for key, names in self.log_names.items():
                 for name in names:
                     self.log_arrays[name][:] = numpy.roll(self.log_arrays[name], -len_diff)
-                    self.log_arrays[name].resize(log_length)
+                    tmparr = numpy.zeros(log_length)
+                    tmparr[:] = self.log_arrays[name][-log_length:]
+                    self.log_arrays[name] = tmparr
+            tmparr = numpy.resize(self.log_xarr, log_length)
+            self.log_xarr = tmparr
+            self.adept_log_xarr()
         self.log_length = log_length
+
+    def adept_log_xarr(self):
+        len = self.log_xarr.size
+        t0 = self.timestamp() - self.backlog
+        t1 = self.timestamp()
+        self.log_xarr = numpy.linspace(t0, t1, len)
+
+    def calc_log_length(self, backlog, frequency):
+        return int(numpy.ceil(backlog * frequency))
+
+    def change_backlog(self, backlog):
+        min_backlog = int(numpy.ceil(self.box_length / self.frequency))
+        if backlog < min_backlog:
+            backlog = min_backlog
+        log_length = self.calc_log_length(backlog, self.frequency)
+        self.backlog = backlog
+        self.change_log_length(log_length)
+
+    def change_frequency(self, frequency):
+        self.frequency = frequency
+        self.change_backlog(self.backlog)
+
 
     def reset_logs(self):
         bc = self.petra.BeamCurrent
@@ -341,6 +396,20 @@ class qbpm():
         for key, names in self.log_names.items():
             for name in names:
                 self.log_arrays[name] = numpy.full(self.log_length, set_vals[name])
+        self.adept_log_xarr()
+
+    def timestamp(self):
+        return time.time()
+
+
+class TimeAxisItem(pg.AxisItem):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.setLabel(text='Time', units=None)
+        self.enableAutoSIPrefix(False)
+
+    def tickStrings(self, values, scale, spacing):
+        return [datetime.datetime.fromtimestamp(value).strftime("%H:%M:%S") for value in values]
 
 
 def main():
